Grid shifting methods implemented by Jonathan Tu. These allow for asymmetric nesting of grids. Grid.h, Grid.cc, and ibpm.cc were modified to implement these methods. Scalar.cc and VectorOperations.cc were also modified, as the indexing must always begin from the left now that the grids are asymmetrically nested. The test files GridTest.cc, ScalarTest.cc, and VectorOperations.cc were also changed to test all methods affected by the new functionality.
A bug was also fixed in Flux.h.  The variable "dir" had been set using integer division before, but this allowed it to take an invalid value of 2 for tall, skinny grids.  This has been changed to an if statement, guaranteeing an acceptable value.

Minor changes were made to the following files to avoid compiler warnings:
    RigidBody.cc, EllipticSover.2d.cc, VectorOperations.cc, RungeKutta3.cc, BC.h, BC.cc, EllipticSolverTest.cc, EllipticSolver2dTest.cc
These changes involved resolving int/double errors, end of file newlines, initializing variables upon declaration, and deleting unused variables.



git-svn-id: https://svn.princeton.edu/rowley/trunk@217 f381375e-2823-4fb2-bad6-22fa55754e8c

+++ src/BC.h
@@ -87,8 +87,8 @@ public:
         return *this;
     }
     
-    inline double Nx() const { return _nx; }
-    inline double Ny() const { return _ny; }
+    inline int Nx() const { return _nx; }
+    inline int Ny() const { return _ny; }
     
 private:
     int _nx;
